netx: unify netx and httpx packages
